[linter] (WIP)
TsLint will be depracated in 2019 (see this blog post
https://medium.com/palantir/tslint-in-2019-1a144c2317a9 and this issue:
https://github.com/palantir/tslint/issues/4534).

This branch attemps to remove it and use ESLint and prettier instead.

This
[doc](https://github.com/typescript-eslint/tslint-to-eslint-config/blob/main/docs/FAQs.md)
and this
[README](https://github.com/typescript-eslint/typescript-eslint/blob/master/docs/getting-started/linting/README.md)
can help through it.

This commit need a fix on the rule `error iterators/generators
require regenerator-runtime, which is too heavyweight
for this guide to allow them. Separately, loops should be
avoided in favor of array iterations no-restricted-syntax`
see
"https://stackoverflow.com/questions/47213651/caught-between-two-no-restricted-syntax-violations"

Please note that one can run eslint with
`npx eslint . --ext .js,.ts,.tsx --fix`